general clean up of dead code and state enums
