+++ lib/blocs/util.dart
@@ -1,0 +1,6 @@
+part of carp_study_app;
+
+extension StringExtension on String {
+  String truncateTo(int maxLength) =>
+      (length <= maxLength) ? this : '${substring(0, maxLength)}...';
+}
